added static method, handle errors, send user data perfectly via zod

+++ src/modules/user/user.controller.ts
@@ -1,13 +1,13 @@
 import { Request, Response } from 'express';
 import { UserServices } from './user.service';
-import { validateUser } from './user.validation';
+import { validateOrder, validateUser } from './user.validation';
+import { UserModel } from '../user.model';
 
 const createAnUser = async (req: Request, res: Response) => {
   try {
     const { user: userData } = req.body;
 
     // Validation using Zod
-
     const zodParsedData = validateUser.parse(userData);
 
     // Calling Service Function to send this data to DB
@@ -16,7 +16,16 @@ const createAnUser = async (req: Request, res: Response) => {
     res.status(200).json({
       success: true,
       message: 'User created successfully!',
-      data: result,
+      data: {
+        userId: result.userId,
+        username: result.username,
+        fullName: result.fullName,
+        age: result.age,
+        email: result.email,
+        isActive: result.isActive,
+        hobbies: result.hobbies,
+        address: result.address,
+      },
     });
   } catch (error) {
     console.error(error);
@@ -27,11 +36,21 @@ const getAllUsers = async (req: Request, res: Response) => {
   try {
     // Calling Service Function to find all users data
     const result = await UserServices.getAllUserFromDB();
+
+    // If no user exists showing no user found
+    if (result.length > 0) {
+      res.status(200).json({
+        success: true,
+        message: 'Users fetched successfully!',
+        data: result,
+      });
+    }
+
     // Sending Response
     res.status(200).json({
       success: true,
-      message: 'Users fetched successfully!',
-      data: result,
+      message: 'No user found.',
+      data: null,
     });
   } catch (error) {
     console.error(error);
@@ -42,6 +61,17 @@ const getSingleUser = async (req: Request, res: Response) => {
   try {
     const { userId } = req.params;
     const numberUserId = parseFloat(userId);
+
+    // Calling User Exists Function from UserModel
+    const userExists = await UserModel.isUserExists(numberUserId);
+    if (!userExists) {
+      return res.status(404).json({
+        success: false,
+        message: 'User not found!',
+        error: { code: 404, description: 'User not found!' },
+      });
+    }
+
     // Calling Service Function to find single user data
     const result = await UserServices.getSingleUserFromDB(numberUserId);
     // Sending Response
@@ -59,6 +89,17 @@ const updateSingleUser = async (req: Request, res: Response) => {
   try {
     const userId = req.params.userId;
     const numberUserId = parseFloat(userId);
+
+    // Calling User Exists Function from UserModel
+    const userExists = await UserModel.isUserExists(numberUserId);
+    if (!userExists) {
+      return res.status(404).json({
+        success: false,
+        message: 'User not found!',
+        error: { code: 404, description: 'User not found!' },
+      });
+    }
+
     const updatedProperties = req.body;
     // Calling Service Function to update single user data
     const result = await UserServices.updateSingleUserFromDB(
@@ -80,6 +121,17 @@ const deleteSingleUser = async (req: Request, res: Response) => {
   try {
     const { userId } = req.params;
     const numberUserId = parseFloat(userId);
+
+    // Calling User Exists Function from UserModel
+    const userExists = await UserModel.isUserExists(numberUserId);
+    if (!userExists) {
+      return res.status(404).json({
+        success: false,
+        message: 'User not found!',
+        error: { code: 404, description: 'User not found!' },
+      });
+    }
+
     // Calling Service Function to delete single user data
     await UserServices.deleteSingleUserFromDB(numberUserId);
     // Sending Response
@@ -97,14 +149,24 @@ const updateSingleUserOrders = async (req: Request, res: Response) => {
   try {
     const userId = req.params.userId;
     const numberUserId = parseFloat(userId);
+
+    // Calling User Exists Function from UserModel
+    const userExists = await UserModel.isUserExists(numberUserId);
+    if (!userExists) {
+      return res.status(404).json({
+        success: false,
+        message: 'User not found!',
+        error: { code: 404, description: 'User not found!' },
+      });
+    }
+
     const updatedProperties = req.body;
     // Zod Parser for orders type
+    const validOrder = validateOrder.parse(updatedProperties);
 
     // Calling Service Function to add orders to single user data
-    await UserServices.updateSingleUseOrdersFromDB(
-      numberUserId,
-      updatedProperties,
-    );
+    await UserServices.updateSingleUseOrdersFromDB(numberUserId, validOrder);
+
     // Sending Response
     res.status(200).json({
       success: true,
@@ -120,8 +182,20 @@ const getOrdersOfSingleUser = async (req: Request, res: Response) => {
   try {
     const { userId } = req.params;
     const numberUserId = parseFloat(userId);
+
+    // Calling User Exists Function from UserModel
+    const userExists = await UserModel.isUserExists(numberUserId);
+    if (!userExists) {
+      return res.status(404).json({
+        success: false,
+        message: 'User not found!',
+        error: { code: 404, description: 'User not found!' },
+      });
+    }
+
     // Calling Service Function to find single user orders data
     const result = await UserServices.getOrdersOfSingleUserFromDB(numberUserId);
+
     // Sending Response
     res.status(200).json({
       success: true,
@@ -137,14 +211,28 @@ const getOrdersTotalPriceOfSingleUser = async (req: Request, res: Response) => {
   try {
     const { userId } = req.params;
     const numberUserId = parseFloat(userId);
+
+    // Calling User Exists Function from UserModel
+    const userExists = await UserModel.isUserExists(numberUserId);
+    if (!userExists) {
+      return res.status(404).json({
+        success: false,
+        message: 'User not found!',
+        error: { code: 404, description: 'User not found!' },
+      });
+    }
+
     // Calling Service Function to find single user's total Price of orders from data
     const result =
       await UserServices.getOrdersTotalPriceOfSingleUserFromDB(numberUserId);
+
     // Sending Response
     res.status(200).json({
       success: true,
       message: 'Total price calculated successfully!',
-      data: result,
+      data: {
+        totalPrice: result.length > 0 ? result[0].totalPrice : 0,
+      },
     });
   } catch (error) {
     console.error(error);
